Extract duplicate code into a helper
